Rename folders.
